home_automation_api/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1
djangorestframework==3.12.1
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
Pillow==7.2.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.26
zipp==3.1.0
